requirements.txt
Django==3.1.3
dj-database-url
django-heroku
gunicorn
psycopg2
python-decouple
whitenoise
pandas
djangorestframework
markdown
django-filter
randomcolor